+++ src/components/__FolderNameList/FolderNameList.module.css
@@ -1,4 +1,4 @@
-.folder-buttons-section {
+/* .folder-buttons-section {
   padding: 2rem 15rem 2rem 4rem;
 }
 
@@ -17,4 +17,4 @@
   font-weight: 600;
   line-height: normal;
   letter-spacing: -0.2px;
-}
+} */
